Avoid unnecessary casts in base::File methods.
The wrapped methods are already returning optional<size_t>. If someone
were to change the checked_casts<> to static_casts<>, we might get
success returned from a partial write of a 4GB+ string, say.

Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7226298
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1554142}

+++ base/files/file.cc
@@ -115,8 +115,7 @@ std::optional<size_t> File::Read(int64_t offset, span<uint8_t> data) {
 }
 
 bool File::ReadAndCheck(int64_t offset, span<uint8_t> data) {
-  // Size checked in span form of Read() above.
-  return Read(offset, data) == static_cast<int>(data.size());
+  return Read(offset, data) == data.size();
 }
 
 std::optional<size_t> File::ReadAtCurrentPos(span<uint8_t> data) {
@@ -131,8 +130,7 @@ std::optional<size_t> File::ReadAtCurrentPos(span<uint8_t> data) {
 }
 
 bool File::ReadAtCurrentPosAndCheck(span<uint8_t> data) {
-  // Size checked in span form of ReadAtCurrentPos() above.
-  return ReadAtCurrentPos(data) == static_cast<int>(data.size());
+  return ReadAtCurrentPos(data) == data.size();
 }
 
 std::optional<size_t> File::Write(int64_t offset, span<const uint8_t> data) {
@@ -147,8 +145,7 @@ std::optional<size_t> File::Write(int64_t offset, span<const uint8_t> data) {
 }
 
 bool File::WriteAndCheck(int64_t offset, span<const uint8_t> data) {
-  // Size checked in span form of Write() above.
-  return Write(offset, data) == static_cast<int>(data.size());
+  return Write(offset, data) == data.size();
 }
 
 std::optional<size_t> File::WriteAtCurrentPos(span<const uint8_t> data) {
@@ -163,8 +160,7 @@ std::optional<size_t> File::WriteAtCurrentPos(span<const uint8_t> data) {
 }
 
 bool File::WriteAtCurrentPosAndCheck(span<const uint8_t> data) {
-  // Size checked in span form of WriteAtCurrentPos() above.
-  return WriteAtCurrentPos(data) == static_cast<int>(data.size());
+  return WriteAtCurrentPos(data) == data.size();
 }
 
 std::optional<size_t> File::ReadAtCurrentPosNoBestEffort(
